abstracted grpc and rabbit mq backend resolver
